исправил недостатки
