+++ js/main.js
@@ -1,5 +1,6 @@
+import './states.js';
 import {renderOffers} from './render.js';
 import './user-form.js';
-import './states.js';
-import './map.js';
 
+
+renderOffers(10);
