+++ js/render.js
@@ -1,5 +1,52 @@
 import {getOffers} from './data.js';
-const similarListElement = document.querySelector('.map__canvas');
+import {turnOnActiveState} from './states.js';
+
+
+const map = L.map('map-canvas')
+  /*.on('load', () => {
+    turnOnActiveState();
+  })*/
+  .setView({
+    lat: 35.70000,
+    lng: 139.42500,
+  }, 10);
+
+L.tileLayer(
+  'https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png',
+  {
+    attribution: '&copy; <a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a> contributors',
+  },
+).addTo(map);
+
+const mainPinIcon = L.icon({
+  iconUrl: '../img/main-pin.svg',
+  iconSize: [52, 52],
+  iconAnchor: [26, 52],
+});
+
+const marker = L.marker(
+  {
+    lat: 35.70000,
+    lng: 139.42500,
+  },
+  {
+    draggable: true,
+    icon: mainPinIcon,
+  },
+);
+
+marker.addTo(map);
+
+marker.on('moveend', (evt) => {
+  address.value = `${evt.target.getLatLng().lat.toFixed(5)  } с.ш. ${  evt.target.getLatLng().lng.toFixed(5)  } в.д.`;
+});
+
+const icon = L.icon({
+  iconUrl: './img/pin.svg',
+  iconSize: [40, 40],
+  iconAnchor: [20, 40],
+});
+
 const similarOfferTemplate = document.querySelector('#card')
   .content
   .querySelector('.popup');
@@ -51,7 +98,6 @@ const renderOffers = (count) => {
     offerElement.querySelector('.popup__text--capacity').textContent = offer.offer.type;
     offerElement.querySelector('.popup__text--capacity').textContent = getDeclension(offer.offer.rooms,offer.offer.guests,['комната', 'комнаты', 'комнат', 'гостя', 'гостей']);
     offerElement.querySelector('.popup__text--time').textContent = `Заезд после ${  offer.offer.checkin  }, выезд до ${  offer.offer.checkout}`;
-
     const featureContainer = offerElement.querySelector('.popup__features');
     const featureListFragment = document.createDocumentFragment();
     offer.offer.features.forEach((featureItem) => {
@@ -77,9 +123,21 @@ const renderOffers = (count) => {
       photoContainer.append(photoListFragment);
     } else {photoContainer.innerHTML = '';}
 
-    similarListElement.appendChild(offerElement);
+    const lat = offer.location.lat;
+    const lng = offer.location.lng;
+    const marker = L.marker(
+      {
+        lat,
+        lng,
+      },
+      {
+        icon,
+      },
+    );
+    marker
+      .addTo(map)
+      .bindPopup(offerElement);
   });
   return similarOffers;
 };
-
 export {renderOffers};
